+++ CHANGELOG.md
@@ -24,6 +24,7 @@ _What's new_
 * Convert credential for mastodon to connection. ([#473](https://github.com/turbot/pipe-fittings/issues/473)).
 * Convert credential for microsoft teams to connection. ([#474](https://github.com/turbot/pipe-fittings/issues/474)).
 * Convert credential for okta to connection. ([#475](https://github.com/turbot/pipe-fittings/issues/475)).
+* Convert credential for openai to connection. ([#476](https://github.com/turbot/pipe-fittings/issues/476)).
 
 _Bug fixes_